Address PR review feedback
- Extract CompactWriter into its own file to reduce OnDiskGraphIndexCompactor size
- Rewrite SystemStatsCollector to read /proc files directly in Java instead of spawning bash
- Clarify recall section description in docs/compaction.md
